+++ src/components/interactive/Button.jsx
@@ -29,7 +29,7 @@ const Button = ({
         'display--inlineFlex',
         'rounded--all',
         'alignChild--center--center',
-        'border--focus--notransition',
+        'border--focus--noTransition',
         'transition-default',
         {
           'fdsButton--s': size === 's',