+++ src/components/interactive/StackedButton.jsx
@@ -14,7 +14,7 @@ const StackedButton = ({ Icon, Link, label, disabled, isActive, hasCaret, ...res
         'fdsStackedButton',
         'rounded--all',
         'fdsStackedButton--ghost',
-        'border--focus--notransition',
+        'border--focus--noTransition',
         'transition-default',
         {
           'fdsStackedButton--disabled': disabled,